+++ moPepGen/util/validate_variant_calling.py
@@ -118,6 +118,8 @@ def call_variant(gvf_files:Path, ref_dir:Path, output_fasta:Path, graph_output_d
     args.threads = 1
     args.timeout_seconds = 900
     args.skip_failed = False
+    args.output_flanking = False
+    args.context_length = 10
     call_variant_peptide(args=args)
 
 def call_brute_force(gvf_files:Path, ref_dir:Path, output_path:str, force:bool,